main index route converted to typescript
